+++ evals/context-query/fixture/src/other.js
@@ -1,0 +1,7 @@
+function slugify(text) {
+  return text;
+}
+
+export function run() {
+  return slugify('a');
+}